src/aasist/components/download_dataset/runtime-requirements.txt
# Generated by KFP.
python-dotenv==1.1.1
requests==2.32.4